v8.0.2011.0929   修正IDataTable.Fix中对一对一关系处理的不足                  修改Entity，根据唯一索引查询单对象时不加分页和排序，这使得在SQLite上特别是MySql上有性能提升

+++ XCode/Entity/IEntityOperate.cs
@@ -52,7 +52,7 @@ namespace XCode
         ISingleEntityCache SingleCache { get; }
 
         /// <summary>总记录数</summary>
-        Int64 Count { get; }
+        Int32 Count { get; }
         #endregion
 
         #region 创建实体
